move _Pytato{Frame,Stack}Summary to tags, convert to dataclass
