+++ pytato/array.py
@@ -1,5 +1,4 @@
 from __future__ import annotations
-from traceback import FrameSummary, StackSummary
 
 __copyright__ = """
 Copyright (C) 2020 Andreas Kloeckner
@@ -1781,58 +1780,6 @@ def _get_default_axes(ndim: int) -> AxesT:
     return tuple(Axis(frozenset()) for _ in range(ndim))
 
 
-@attrs.define(frozen=True, eq=True)
-class _PytatoFrameSummary:
-    """Class to store a single call frame, similar to
-    :class:`FrameSummary`, but immutable."""
-    filename: str
-    lineno: Optional[int]
-    name: str
-    line: Optional[str]
-
-    def short_str(self, maxlen: int = 100) -> str:
-        s = f"{self.filename}:{self.lineno}, in {self.name}():\n{self.line}"
-        s1, s2 = s.split("\n")
-        # Limit display to maxlen characters
-        s1 = "[...] " + s1[len(s1)-maxlen:] if len(s1) > maxlen else s1
-        s2 = s2[:maxlen] + " [...]" if len(s2) > maxlen else s2
-        return s1 + "\n" + s2
-
-    def __repr__(self) -> str:
-        return f"{self.filename}:{self.lineno}, in {self.name}(): {self.line}"
-
-
-@attrs.define(frozen=True, eq=True)
-class _PytatoStackSummary:
-    """Class to store a list of :class:`_PytatoFrameSummary` call frames,
-    similar to :class:`StackSummary`, but immutable."""
-    frames: Tuple[_PytatoFrameSummary, ...]
-
-    def to_stacksummary(self) -> StackSummary:
-        frames = [FrameSummary(f.filename, f.lineno, f.name, line=f.line)
-                  for f in self.frames]
-
-        return StackSummary.from_list(frames)
-
-    def short_str(self, maxlen: int = 100) -> str:
-        from os.path import dirname
-
-        # Find the first file in the frames that is not in pytato's internal
-        # directories.
-        for frame in reversed(self.frames):
-            frame_dir = dirname(frame.filename)
-            if (not frame_dir.endswith("pytato")
-                    and not frame_dir.endswith("pytato/distributed")):
-                return frame.short_str(maxlen)
-
-        # Fallback in case we don't find any file that is not in the pytato/
-        # directory (should be unlikely).
-        return self.__repr__()
-
-    def __repr__(self) -> str:
-        return "\n  " + "\n  ".join([str(f) for f in self.frames])
-
-
 _ENABLE_TRACEBACK_TAG = False
 
 
@@ -1858,6 +1805,8 @@ def _get_created_at_tag(stacklevel: int = 1) -> Optional[Tag]:
     # Drop the stack levels corresponding to extract_stack() and any additional
     # levels specified via stacklevel
     stack = traceback.extract_stack()[:-(1+stacklevel)]
+
+    from pytato.tags import _PytatoFrameSummary, _PytatoStackSummary
 
     frames = tuple(_PytatoFrameSummary(s.filename, s.lineno, s.name, s.line)
                        for s in stack)
